src/requirements/local.txt
# Extend base.txt
-r test.txt

# Add local extra requirements here (django-debug etc)
black
django-debug-toolbar
isort
pipdeptree
ptvsd
django-stubs>=1.12,<1.13
djangorestframework-stubs>=1.7,<1.8
django-extensions
